Added cacheing

+++ src/main/java/org/example/qposbackend/Stock/stocktaking/stocktakeItem/StockTakeItemDto.java
@@ -1,3 +1,3 @@
 package org.example.qposbackend.Stock.stocktaking.stocktakeItem;
 
-public record StockTakeItemDto(Long id, Integer quantity) {}
+public record StockTakeItemDto(Long id, Double quantity) {}
